report_db/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-autoslug==1.9.7
django-bootstrap4==1.1.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-import-export==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
Faker==4.0.0
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
misaka==2.1.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.19
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tablib==2.0.0
text-unidecode==1.3
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
